requirements.txt
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.7.0"
gitdb==4.0.10; python_version >= "3.7"
gitpython==3.1.31; python_version >= "3.7"
smmap==5.0.0; python_version >= "3.7"
typer==0.9.0; python_version >= "3.6"
typing-extensions==4.6.3; python_version >= "3.7"
